# Reduced garbage produced on label lookup