@brian-brazil 2018-12-13
Can you share your benchmark results? I'm particularly interested if the single label case optimistion is worth it.