@franz1981 2018-12-13
Yes, sure!
I just have the case with 2 label names ie pooled/not pooled that is quite similar.
Wait a sec and I will post it :+1: