@franz1981 2018-12-13
On master:
```
Benchmark                                                                                            Mode  Samples    Score    Error   Units
i.p.b.SummaryBenchmark.prometheusSimpleHistogramBenchmark                                            avgt        4   42.288 ±  8.848   ns/op
i.p.b.SummaryBenchmark.prometheusSimpleHistogramBenchmark:@gc.count.profiled                         avgt        4    5.000 ±    NaN  counts
i.p.b.SummaryBenchmark.prometheusSimpleHistogramBenchmark:@gc.count.total                            avgt        4   18.000 ±    NaN  counts
i.p.b.SummaryBenchmark.prometheusSimpleHistogramBenchmark:@gc.time.profiled                          avgt        4    7.000 ±    NaN      ms
i.p.b.SummaryBenchmark.prometheusSimpleHistogramBenchmark:@gc.time.total                             avgt        4   52.000 ±    NaN      ms
i.p.b.SummaryBenchmark.prometheusSimpleHistogramPooledLabelNamesBenchmark                            avgt        4   45.019 ± 18.353   ns/op
i.p.b.SummaryBenchmark.prometheusSimpleHistogramPooledLabelNamesBenchmark:@gc.count.profiled         avgt        4   19.000 ±    NaN  counts
i.p.b.SummaryBenchmark.prometheusSimpleHistogramPooledLabelNamesBenchmark:@gc.count.total            avgt        4   37.000 ±    NaN  counts
i.p.b.SummaryBenchmark.prometheusSimpleHistogramPooledLabelNamesBenchmark:@gc.time.profiled          avgt        4   21.000 ±    NaN      ms
i.p.b.SummaryBenchmark.prometheusSimpleHistogramPooledLabelNamesBenchmark:@gc.time.total             avgt        4   51.000 ±    NaN      ms
i.p.b.SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark                                       avgt        4  112.427 ± 58.295   ns/op
i.p.b.SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:@gc.count.profiled                    avgt        4    9.000 ±    NaN  counts
i.p.b.SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:@gc.count.total                       avgt        4   23.000 ±    NaN  counts
i.p.b.SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:@gc.time.profiled                     avgt        4   11.000 ±    NaN      ms
i.p.b.SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:@gc.time.total                        avgt        4   41.000 ±    NaN      ms
i.p.b.SummaryBenchmark.prometheusSimpleHistogramTimerPooledLabelNamesBenchmark                       avgt        4   90.238 ± 18.973   ns/op
i.p.b.SummaryBenchmark.prometheusSimpleHistogramTimerPooledLabelNamesBenchmark:@gc.count.profiled    avgt        4    5.000 ±    NaN  counts
i.p.b.SummaryBenchmark.prometheusSimpleHistogramTimerPooledLabelNamesBenchmark:@gc.count.total       avgt        4   16.000 ±    NaN  counts
i.p.b.SummaryBenchmark.prometheusSimpleHistogramTimerPooledLabelNamesBenchmark:@gc.time.profiled     avgt        4    5.000 ±    NaN      ms
i.p.b.SummaryBenchmark.prometheusSimpleHistogramTimerPooledLabelNamesBenchmark:@gc.time.total        avgt        4   26.000 ±    NaN      ms
i.p.b.SummaryBenchmark.prometheusSimpleSummaryBenchmark                                              avgt        4   39.179 ±  5.010   ns/op
i.p.b.SummaryBenchmark.prometheusSimpleSummaryBenchmark:@gc.count.profiled                           avgt        4   30.000 ±    NaN  counts
i.p.b.SummaryBenchmark.prometheusSimpleSummaryBenchmark:@gc.count.total                              avgt        4   52.000 ±    NaN  counts
i.p.b.SummaryBenchmark.prometheusSimpleSummaryBenchmark:@gc.time.profiled                            avgt        4   30.000 ±    NaN      ms
i.p.b.SummaryBenchmark.prometheusSimpleSummaryBenchmark:@gc.time.total                               avgt        4   57.000 ±    NaN      ms
i.p.b.SummaryBenchmark.prometheusSimpleSummaryPooledLabelNamesBenchmark                              avgt        4   38.892 ±  2.868   ns/op
i.p.b.SummaryBenchmark.prometheusSimpleSummaryPooledLabelNamesBenchmark:@gc.count.profiled           avgt        4   26.000 ±    NaN  counts
i.p.b.SummaryBenchmark.prometheusSimpleSummaryPooledLabelNamesBenchmark:@gc.count.total              avgt        4   45.000 ±    NaN  counts
i.p.b.SummaryBenchmark.prometheusSimpleSummaryPooledLabelNamesBenchmark:@gc.time.profiled            avgt        4   27.000 ±    NaN      ms
i.p.b.SummaryBenchmark.prometheusSimpleSummaryPooledLabelNamesBenchmark:@gc.time.total               avgt        4   61.000 ±    NaN      ms
```