@franz1981 2018-12-14
MASTER:
```
Benchmark                                                                                                  Mode  Cnt     Score     Error   Units
SummaryBenchmark.prometheusSimpleHistogramBenchmark                                                        avgt    4    44.351 ±  15.006   ns/op
SummaryBenchmark.prometheusSimpleHistogramBenchmark:·gc.alloc.rate                                         avgt    4   984.933 ± 327.968  MB/sec
SummaryBenchmark.prometheusSimpleHistogramBenchmark:·gc.alloc.rate.norm                                    avgt    4    48.000 ±   0.001    B/op
SummaryBenchmark.prometheusSimpleHistogramBenchmark:·gc.churn.PS_Eden_Space                                avgt    4   984.451 ± 248.087  MB/sec
SummaryBenchmark.prometheusSimpleHistogramBenchmark:·gc.churn.PS_Eden_Space.norm                           avgt    4    48.000 ±   3.983    B/op
SummaryBenchmark.prometheusSimpleHistogramBenchmark:·gc.churn.PS_Survivor_Space                            avgt    4     0.069 ±   0.071  MB/sec
SummaryBenchmark.prometheusSimpleHistogramBenchmark:·gc.churn.PS_Survivor_Space.norm                       avgt    4     0.003 ±   0.003    B/op
SummaryBenchmark.prometheusSimpleHistogramBenchmark:·gc.count                                              avgt    4   276.000            counts
SummaryBenchmark.prometheusSimpleHistogramBenchmark:·gc.time                                               avgt    4   318.000                ms
SummaryBenchmark.prometheusSimpleHistogramPooledLabelNamesBenchmark                                        avgt    4    45.660 ±  11.235   ns/op
SummaryBenchmark.prometheusSimpleHistogramPooledLabelNamesBenchmark:·gc.alloc.rate                         avgt    4   477.863 ± 114.849  MB/sec
SummaryBenchmark.prometheusSimpleHistogramPooledLabelNamesBenchmark:·gc.alloc.rate.norm                    avgt    4    24.000 ±   0.001    B/op
SummaryBenchmark.prometheusSimpleHistogramPooledLabelNamesBenchmark:·gc.churn.PS_Eden_Space                avgt    4   476.789 ± 105.479  MB/sec
SummaryBenchmark.prometheusSimpleHistogramPooledLabelNamesBenchmark:·gc.churn.PS_Eden_Space.norm           avgt    4    23.949 ±   1.579    B/op
SummaryBenchmark.prometheusSimpleHistogramPooledLabelNamesBenchmark:·gc.churn.PS_Survivor_Space            avgt    4     0.015 ±   0.042  MB/sec
SummaryBenchmark.prometheusSimpleHistogramPooledLabelNamesBenchmark:·gc.churn.PS_Survivor_Space.norm       avgt    4     0.001 ±   0.002    B/op
SummaryBenchmark.prometheusSimpleHistogramPooledLabelNamesBenchmark:·gc.count                              avgt    4   262.000            counts
SummaryBenchmark.prometheusSimpleHistogramPooledLabelNamesBenchmark:·gc.time                               avgt    4   308.000                ms
SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark                                                   avgt    4    86.875 ±   2.624   ns/op
SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:·gc.alloc.rate                                    avgt    4   501.808 ±  15.102  MB/sec
SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:·gc.alloc.rate.norm                               avgt    4    48.000 ±   0.001    B/op
SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:·gc.churn.PS_Eden_Space                           avgt    4   502.020 ±  17.032  MB/sec
SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:·gc.churn.PS_Eden_Space.norm                      avgt    4    48.021 ±   1.611    B/op
SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:·gc.churn.PS_Survivor_Space                       avgt    4     0.061 ±   0.110  MB/sec
SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:·gc.churn.PS_Survivor_Space.norm                  avgt    4     0.006 ±   0.011    B/op
SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:·gc.count                                         avgt    4   324.000            counts
SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:·gc.time                                          avgt    4   311.000                ms
SummaryBenchmark.prometheusSimpleHistogramTimerPooledLabelNamesBenchmark                                   avgt    4    85.491 ±   0.773   ns/op
SummaryBenchmark.prometheusSimpleHistogramTimerPooledLabelNamesBenchmark:·gc.alloc.rate                    avgt    4   254.962 ±   2.280  MB/sec
SummaryBenchmark.prometheusSimpleHistogramTimerPooledLabelNamesBenchmark:·gc.alloc.rate.norm               avgt    4    24.000 ±   0.001    B/op
SummaryBenchmark.prometheusSimpleHistogramTimerPooledLabelNamesBenchmark:·gc.churn.PS_Eden_Space           avgt    4   254.942 ±   3.027  MB/sec
SummaryBenchmark.prometheusSimpleHistogramTimerPooledLabelNamesBenchmark:·gc.churn.PS_Eden_Space.norm      avgt    4    23.998 ±   0.341    B/op
SummaryBenchmark.prometheusSimpleHistogramTimerPooledLabelNamesBenchmark:·gc.churn.PS_Survivor_Space       avgt    4     0.017 ±   0.060  MB/sec
SummaryBenchmark.prometheusSimpleHistogramTimerPooledLabelNamesBenchmark:·gc.churn.PS_Survivor_Space.norm  avgt    4     0.002 ±   0.006    B/op
SummaryBenchmark.prometheusSimpleHistogramTimerPooledLabelNamesBenchmark:·gc.count                         avgt    4   298.000            counts
SummaryBenchmark.prometheusSimpleHistogramTimerPooledLabelNamesBenchmark:·gc.time                          avgt    4   309.000                ms
SummaryBenchmark.prometheusSimpleSummaryBenchmark                                                          avgt    4    38.521 ±   6.949   ns/op
SummaryBenchmark.prometheusSimpleSummaryBenchmark:·gc.alloc.rate                                           avgt    4  1132.375 ± 204.787  MB/sec
SummaryBenchmark.prometheusSimpleSummaryBenchmark:·gc.alloc.rate.norm                                      avgt    4    48.000 ±   0.001    B/op
SummaryBenchmark.prometheusSimpleSummaryBenchmark:·gc.churn.PS_Eden_Space                                  avgt    4  1131.893 ± 184.905  MB/sec
SummaryBenchmark.prometheusSimpleSummaryBenchmark:·gc.churn.PS_Eden_Space.norm                             avgt    4    47.982 ±   0.990    B/op
SummaryBenchmark.prometheusSimpleSummaryBenchmark:·gc.churn.PS_Survivor_Space                              avgt    4     0.077 ±   0.029  MB/sec
SummaryBenchmark.prometheusSimpleSummaryBenchmark:·gc.churn.PS_Survivor_Space.norm                         avgt    4     0.003 ±   0.001    B/op
SummaryBenchmark.prometheusSimpleSummaryBenchmark:·gc.count                                                avgt    4   339.000            counts
SummaryBenchmark.prometheusSimpleSummaryBenchmark:·gc.time                                                 avgt    4   328.000                ms
SummaryBenchmark.prometheusSimpleSummaryPooledLabelNamesBenchmark                                          avgt    4    38.991 ±   9.484   ns/op
SummaryBenchmark.prometheusSimpleSummaryPooledLabelNamesBenchmark:·gc.alloc.rate                           avgt    4   559.594 ± 131.568  MB/sec
SummaryBenchmark.prometheusSimpleSummaryPooledLabelNamesBenchmark:·gc.alloc.rate.norm                      avgt    4    24.000 ±   0.001    B/op
SummaryBenchmark.prometheusSimpleSummaryPooledLabelNamesBenchmark:·gc.churn.PS_Eden_Space                  avgt    4   558.727 ± 135.295  MB/sec
SummaryBenchmark.prometheusSimpleSummaryPooledLabelNamesBenchmark:·gc.churn.PS_Eden_Space.norm             avgt    4    23.962 ±   0.852    B/op
SummaryBenchmark.prometheusSimpleSummaryPooledLabelNamesBenchmark:·gc.churn.PS_Survivor_Space              avgt    4     0.014 ±   0.033  MB/sec
SummaryBenchmark.prometheusSimpleSummaryPooledLabelNamesBenchmark:·gc.churn.PS_Survivor_Space.norm         avgt    4     0.001 ±   0.001    B/op
SummaryBenchmark.prometheusSimpleSummaryPooledLabelNamesBenchmark:·gc.count                                avgt    4   333.000            counts
SummaryBenchmark.prometheusSimpleSummaryPooledLabelNamesBenchmark:·gc.time                                 avgt    4   335.000                ms
```
THIS PR:
```
PR:

Benchmark                                                                                     Mode  Cnt     Score      Error   Units
SummaryBenchmark.prometheusSimpleHistogramBenchmark                                           avgt    4    44.648 ±    1.256   ns/op
SummaryBenchmark.prometheusSimpleHistogramBenchmark:·gc.alloc.rate                            avgt    4   488.204 ±   13.789  MB/sec
SummaryBenchmark.prometheusSimpleHistogramBenchmark:·gc.alloc.rate.norm                       avgt    4    24.000 ±    0.001    B/op
SummaryBenchmark.prometheusSimpleHistogramBenchmark:·gc.churn.PS_Eden_Space                   avgt    4   488.006 ±   27.472  MB/sec
SummaryBenchmark.prometheusSimpleHistogramBenchmark:·gc.churn.PS_Eden_Space.norm              avgt    4    23.990 ±    1.099    B/op
SummaryBenchmark.prometheusSimpleHistogramBenchmark:·gc.churn.PS_Survivor_Space               avgt    4     0.015 ±    0.042  MB/sec
SummaryBenchmark.prometheusSimpleHistogramBenchmark:·gc.churn.PS_Survivor_Space.norm          avgt    4     0.001 ±    0.002    B/op
SummaryBenchmark.prometheusSimpleHistogramBenchmark:·gc.count                                 avgt    4   349.000             counts
SummaryBenchmark.prometheusSimpleHistogramBenchmark:·gc.time                                  avgt    4   341.000                 ms
SummaryBenchmark.prometheusSimpleHistogramPooledLabelNamesBenchmark                           avgt    4    44.826 ±    0.786   ns/op
SummaryBenchmark.prometheusSimpleHistogramPooledLabelNamesBenchmark:·gc.alloc.rate            avgt    4    ≈ 10⁻⁴             MB/sec
SummaryBenchmark.prometheusSimpleHistogramPooledLabelNamesBenchmark:·gc.alloc.rate.norm       avgt    4    ≈ 10⁻⁶               B/op
SummaryBenchmark.prometheusSimpleHistogramPooledLabelNamesBenchmark:·gc.count                 avgt    4       ≈ 0             counts
SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark                                      avgt    4    88.715 ±    1.072   ns/op
SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:·gc.alloc.rate                       avgt    4   245.693 ±    2.951  MB/sec
SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:·gc.alloc.rate.norm                  avgt    4    24.000 ±    0.001    B/op
SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:·gc.churn.PS_Eden_Space              avgt    4   245.905 ±    2.749  MB/sec
SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:·gc.churn.PS_Eden_Space.norm         avgt    4    24.021 ±    0.215    B/op
SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:·gc.churn.PS_Survivor_Space          avgt    4     0.016 ±    0.066  MB/sec
SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:·gc.churn.PS_Survivor_Space.norm     avgt    4     0.002 ±    0.006    B/op
SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:·gc.count                            avgt    4   343.000             counts
SummaryBenchmark.prometheusSimpleHistogramTimerBenchmark:·gc.time                             avgt    4   327.000                 ms
SummaryBenchmark.prometheusSimpleHistogramTimerPooledLabelNamesBenchmark                      avgt    4    90.161 ±    2.881   ns/op
SummaryBenchmark.prometheusSimpleHistogramTimerPooledLabelNamesBenchmark:·gc.alloc.rate       avgt    4    ≈ 10⁻⁴             MB/sec
SummaryBenchmark.prometheusSimpleHistogramTimerPooledLabelNamesBenchmark:·gc.alloc.rate.norm  avgt    4    ≈ 10⁻⁵               B/op
SummaryBenchmark.prometheusSimpleHistogramTimerPooledLabelNamesBenchmark:·gc.count            avgt    4       ≈ 0             counts
SummaryBenchmark.prometheusSimpleSummaryBenchmark                                             avgt    4    43.756 ±    1.004   ns/op
SummaryBenchmark.prometheusSimpleSummaryBenchmark:·gc.alloc.rate                              avgt    4   498.160 ±   11.422  MB/sec
SummaryBenchmark.prometheusSimpleSummaryBenchmark:·gc.alloc.rate.norm                         avgt    4    24.000 ±    0.001    B/op
SummaryBenchmark.prometheusSimpleSummaryBenchmark:·gc.churn.PS_Eden_Space                     avgt    4   497.915 ±   53.488  MB/sec
SummaryBenchmark.prometheusSimpleSummaryBenchmark:·gc.churn.PS_Eden_Space.norm                avgt    4    23.988 ±    2.172    B/op
SummaryBenchmark.prometheusSimpleSummaryBenchmark:·gc.churn.PS_Survivor_Space                 avgt    4     0.018 ±    0.079  MB/sec
SummaryBenchmark.prometheusSimpleSummaryBenchmark:·gc.churn.PS_Survivor_Space.norm            avgt    4     0.001 ±    0.004    B/op
SummaryBenchmark.prometheusSimpleSummaryBenchmark:·gc.count                                   avgt    4   273.000             counts
SummaryBenchmark.prometheusSimpleSummaryBenchmark:·gc.time                                    avgt    4   297.000                 ms
SummaryBenchmark.prometheusSimpleSummaryPooledLabelNamesBenchmark                             avgt    4    42.355 ±    3.602   ns/op
SummaryBenchmark.prometheusSimpleSummaryPooledLabelNamesBenchmark:·gc.alloc.rate              avgt    4    ≈ 10⁻⁴             MB/sec
SummaryBenchmark.prometheusSimpleSummaryPooledLabelNamesBenchmark:·gc.alloc.rate.norm         avgt    4    ≈ 10⁻⁶               B/op
SummaryBenchmark.prometheusSimpleSummaryPooledLabelNamesBenchmark:·gc.count                   avgt    4       ≈ 0             counts
```
Looking at the number and thanks to the changes to speed-up `ArrayList:.equals/hashCode` the performance are similar (often a lil' worst TBH) to the original version, but with cached labels names (`String[]`) it is possible to achieve 0 garbage on the hot path, that help to avoid latency spikes on the rest of application that use it.
I will come soon with a bench that cover the single label case too :+1: