@franz1981 2018-12-14
While Re label names lookup perf:
On Master:
```
Benchmark                                                                             (labelNamesCount)  Mode  Cnt     Score     Error   Units
LabelNamesLookupBenchmark.labelNamesLookupBenchmark                                                   1  avgt    8    27.037 ±   2.878   ns/op
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.alloc.rate                                    1  avgt    8  1616.415 ± 156.981  MB/sec
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.alloc.rate.norm                               1  avgt    8    48.000 ±   0.001    B/op
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.churn.PS_Eden_Space                           1  avgt    8  1616.002 ± 149.652  MB/sec
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.churn.PS_Eden_Space.norm                      1  avgt    8    47.993 ±   0.352    B/op
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.churn.PS_Survivor_Space                       1  avgt    8     0.071 ±   0.034  MB/sec
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.churn.PS_Survivor_Space.norm                  1  avgt    8     0.002 ±   0.001    B/op
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.count                                         1  avgt    8   614.000            counts
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.time                                          1  avgt    8   639.000                ms
LabelNamesLookupBenchmark.labelNamesLookupBenchmark                                                   2  avgt    8    29.165 ±   3.770   ns/op
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.alloc.rate                                    2  avgt    8   750.260 ±  93.636  MB/sec
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.alloc.rate.norm                               2  avgt    8    24.000 ±   0.001    B/op
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.churn.PS_Eden_Space                           2  avgt    8   748.556 ±  91.635  MB/sec
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.churn.PS_Eden_Space.norm                      2  avgt    8    23.948 ±   0.457    B/op
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.churn.PS_Survivor_Space                       2  avgt    8     0.018 ±   0.011  MB/sec
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.churn.PS_Survivor_Space.norm                  2  avgt    8     0.001 ±   0.001    B/op
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.count                                         2  avgt    8   632.000            counts
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.time                                          2  avgt    8   622.000                ms
```
With this PR:
```
Benchmark                                                                (labelNamesCount)  Mode  Cnt   Score    Error   Units
LabelNamesLookupBenchmark.labelNamesLookupBenchmark                                      1  avgt    8  25.569 ±  2.521   ns/op
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.alloc.rate                       1  avgt    8  ≈ 10⁻⁴           MB/sec
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.alloc.rate.norm                  1  avgt    8  ≈ 10⁻⁶             B/op
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.count                            1  avgt    8     ≈ 0           counts
LabelNamesLookupBenchmark.labelNamesLookupBenchmark                                      2  avgt    8  41.671 ±  4.040   ns/op
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.alloc.rate                       2  avgt    8  ≈ 10⁻⁴           MB/sec
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.alloc.rate.norm                  2  avgt    8  ≈ 10⁻⁶             B/op
LabelNamesLookupBenchmark.labelNamesLookupBenchmark:·gc.count                            2  avgt    8     ≈ 0           counts
```
We just pay more cost CPU-wise on the label names count = 2, but much less GC (0), while with a single label name the perf are improved with no GC at all